Moved intersect into a new package and added related unit test.
