+++ pkg/bot/teams.go
@@ -21,7 +21,7 @@ import (
 	"github.com/kubeshop/botkube/pkg/format"
 	"github.com/kubeshop/botkube/pkg/httpsrv"
 	"github.com/kubeshop/botkube/pkg/multierror"
-	"github.com/kubeshop/botkube/pkg/utils"
+	"github.com/kubeshop/botkube/pkg/sliceutil"
 )
 
 // TODO: Refactor this file as a part of https://github.com/kubeshop/botkube/issues/667
@@ -317,7 +317,7 @@ func (b *Teams) SendEvent(ctx context.Context, event events.Event, eventSources 
 	b.log.Debugf(">> Sending to Teams: %+v", event)
 	card := b.formatMessage(event, b.Notification)
 
-	if !utils.Intersect(eventSources, b.bindings.Sources) {
+	if !sliceutil.Intersect(eventSources, b.bindings.Sources) {
 		b.log.Debugf(
 			"Event was not sent as bot source bindings: %+v do not overlap with the event's sources: %+v",
 			b.bindings.Sources,
