+++ pkg/sliceutil/intersect.go
@@ -1,4 +1,4 @@
-package utils
+package sliceutil
 
 import (
 	"strings"
